Client : Refactor adjust filters pathing. Fix bug where settingsnavigation pushes edit forms off the page
