+++ client/src/components/views/seb-server/settings-navigation/components/SettingsTable/SettingsTable.vue
@@ -127,8 +127,8 @@
 <script setup lang="ts">
 import { computed, ref } from "vue";
 import type { SettingsTableHeader } from "@/components/views/seb-server/settings-navigation/components/SettingsTable/types.ts";
-import DeleteDialog from "@/components/views/seb-server/settings-navigation/components/DeleteDialog.vue";
-import StatusDialog from "@/components/views/seb-server/settings-navigation/components/StatusDialog.vue";
+import DeleteDialog from "@/components/views/seb-server/settings-navigation/components/SettingsTable/widgets/DeleteDialog.vue";
+import StatusDialog from "@/components/views/seb-server/settings-navigation/components/SettingsTable/widgets/StatusDialog.vue";
 import { useSettingsNavigation } from "@/components/views/seb-server/settings-navigation/components/SettingsTable/composables/useSettingsNavigation.ts";
 import type { ServerTablePaging } from "@/models/types";
 import type { RouteName } from "@/router/routeNames";
