+++ client/src/components/views/seb-server/settings-navigation/components/SettingsTable/widgets/DeleteDialog.vue
